requirements.txt
flask
google-generativeai
gTTS
pypdf
python-dotenv
